refactor(extract): isolate case-name scanner
